+++ field/field-locator-contact.js
@@ -3,6 +3,7 @@ const FieldLocator = require('./field-locator').FieldLocator;
 const FieldTextBoolean = require('./field-text-boolean').FieldTextBoolean;
 const FieldLocatorText = require('./field-locator').FieldLocatorText;
 const FieldLocatorGuid = require('./field-locator').FieldLocatorGuid;
+const NameParse = require('../lib/name-parser').ParseFullName;
 const _ = require('lodash');
 
 class FieldLocatorContact extends FieldLocator {
@@ -12,6 +13,7 @@ class FieldLocatorContact extends FieldLocator {
     this._fields = _.merge(this._fields, {
       id: new FieldLocatorGuid(),
       fullName: new FieldLocatorText(),
+      trueName: new FieldLocatorText(),
       firstName: new FieldLocatorText(),
       namePrefix: new FieldLocatorText(),
       name: new FieldLocatorGuid(),
@@ -26,6 +28,7 @@ class FieldLocatorContact extends FieldLocator {
       _allowMulti: new FieldTextBoolean()
     });
     this._fields.organisation = this._fields.name;
+    this.addStoreGroup('trueName');
     this.addStoreGroup('fullName');
     this.addStoreGroup('name');
     this.addStoreGroup('guid');
@@ -40,6 +43,31 @@ class FieldLocatorContact extends FieldLocator {
     if (data.organisation) {
       data.name = data.organisation
       fields.name = fields.organisation
+    } else if (data.trueName) {
+      // split Erick de Boer into fullName: Boer, Erick de
+      let parser = new NameParse();
+      let parsed = parser.analyse(data.trueName);
+      if (parsed.error.length) {
+        this.log(logger, 'warn', fieldName + 'trueName', parsed.error.join(', '));
+      }
+      const mapping = {
+        last: 'name',
+        first: 'firstName',
+        middle: 'middleName',
+        nick: 'nickName',
+        // what to do with the middle name and nick
+        title: 'title',
+        prefix: 'namePrefix',
+        suffix: 'nameSuffix'
+      };
+      let foundData = {};
+      for (let field in mapping) {
+        if (parsed[field].length) {
+          foundData[mapping[field]] = parsed[field].trim();
+        }
+      }
+      delete data.trueName;
+      data.fullName = foundData.name + ', ' + foundData.firstName + ( foundData.middleName? ` ${foundData.middleName}` : '') + (foundData.namePrefix ? ` ${foundData.namePrefix}` : '');
     }
     return super.processKeys(fieldName, fields, data, logger)
   }